+++ src/PDE/Integrate/Boundary.cpp
@@ -103,7 +103,7 @@ bndSurfInt( std::size_t nmat,
   //        "derivative vector for single material compflow" );
 
   for (const auto& s : bcconfig) {       // for all bc sidesets
-    auto bc = bface.find(s);// faces for side set
+    auto bc = bface.find(static_cast<int>(s));// faces for side set
     if (bc != end(bface))
     {
       for (const auto& f : bc->second)
@@ -362,7 +362,7 @@ bndSurfIntFV(
   auto nprim = P.nprop()/rdof;
 
   for (const auto& s : bcconfig) {       // for all bc sidesets
-    auto bc = bface.find(s);// faces for side set
+    auto bc = bface.find(static_cast<int>(s));// faces for side set
     if (bc != end(bface))
     {
       for (const auto& f : bc->second)
